Move overlap implementation into Overlap companion.
